Move util functions to separate file
