+++ serverless/lib/index.js
@@ -1,20 +1,6 @@
 var request = require('request');
 var Promise = require('bluebird');
 
-module.exports.parseFormData = function(data) {
-  return data.split(', ').reduce(function(object, line) {
-    var variables = line.split('=');
-    object[variables[0]] = variables[1];
-    return object;
-  }, {});
-};
-
-module.exports.mapToArray = function(map) {
-  return Object.keys(map).map(function(key) {
-    return map[key];
-  });
-};
-
 module.exports.injectTeamsToGames = function(teamMap, games) {
   return games.map(function(game) {
     game.team1 = [teamMap[game.team1].player1, teamMap[game.team1].player2];
